change double data type to BigDecimal for currency, modify db data types (#3)

+++ src/java/com/allan/amca/transaction/Transactional.java
@@ -1,5 +1,7 @@
 package com.allan.amca.transaction;
 
+import java.math.BigDecimal;
+
 /**
  * @author allanaranzaso
  */
@@ -10,5 +12,5 @@ public interface Transactional {
      * @param amount the amount to either withdraw or deposit.
      * @return true if the query executed successfully. Otherwise, false.
      */
-    boolean performTransaction(long client, Double amount);
+    boolean performTransaction(long client, BigDecimal amount);
 }
